+++ src/services/BaseService.php
@@ -1,0 +1,10 @@
+<?php
+namespace wheelform\services;
+
+use craft\helpers\Template as TemplateHelper;
+use yii\base\BaseObject;
+
+abstract class BaseService extends BaseObject
+{
+
+}